ecomm/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-admin-thumbnails==0.2.8
django-session-timeout==0.1.0
frozenlist==1.4.1
idna==3.6
kombu==5.3.6
multidict==6.0.5
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==9.0.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
